fixed64: codemod pass 4 -- Fixed->scalar casts in #included files
The per-TU codemod only fixed errors in the compiled file, not in
#included sources (a_*misc.cpp etc.). Applied the same behavior-
preserving Fixed->scalar casts to those from the tree-wide error dump:
66 casts across 26 files (assigning-to / no-viable-conversion narrowings,
range = the RHS expression, cast to the exact target type -- identical
truncation in both modes). These are mechanical conversions, not
computations, so they carry no unit test; the flag-off build (green) and
the planned flag-off-vs-flag-on differential are their correctness gate.
Genuine computations remain extracted+tested in features/*/computation.
Tests green.

+++ src/zandronum/src/g_doom/a_lostsoul.cpp
@@ -49,7 +49,7 @@ void A_SkullAttack(AActor *self, fixed_t speed)
 	an = self->angle >> ANGLETOFINESHIFT;
 	self->velx = FixedMul (speed, finecosine[an]);
 	self->vely = FixedMul (speed, finesine[an]);
-	dist = P_AproxDistance (dest->x - self->x, dest->y - self->y);
+	dist = (int)(P_AproxDistance (dest->x - self->x, dest->y - self->y));
 	dist = dist / speed;
 	
 	if (dist < 1)
