+++ src/zandronum/src/g_shared/sbarinfo_commands.cpp
@@ -98,7 +98,7 @@ class CommandDrawImage : public SBarInfoCommandFlowControl
 				w=(int) (texture->GetScaledWidthDouble()*spawnScaleX);
 				h=(int) (texture->GetScaledHeightDouble()*spawnScaleY);
 			}
-			statusBar->DrawGraphic(texture, imgx, imgy, block->XOffset(), block->YOffset(), frameAlpha, block->FullScreenOffsets(),
+			statusBar->DrawGraphic(texture, imgx, imgy, block->XOffset(), block->YOffset(), (int)(frameAlpha), block->FullScreenOffsets(),
 				translatable, false, offset, false, w, h);
 		}
 		void	Parse(FScanner &sc, bool fullScreenOffsets)
@@ -243,7 +243,7 @@ class CommandDrawImage : public SBarInfoCommandFlowControl
 				return;
 
 			texture = NULL;
-			alpha = FRACUNIT;
+			alpha = (int)(FRACUNIT);
 			if (applyscale)
 			{
 				spawnScaleX = spawnScaleY = 1.0f;
@@ -291,7 +291,7 @@ class CommandDrawImage : public SBarInfoCommandFlowControl
 					if (harmor->Slots[armorType] > 0 && harmor->SlotsIncrement[armorType] > 0)
 					{
 						//combine the alpha values
-						alpha = fixed_t(((double) alpha / (double) FRACUNIT) * ((double) MIN<fixed_t> (OPAQUE, Scale(harmor->Slots[armorType], OPAQUE, harmor->SlotsIncrement[armorType])) / (double) OPAQUE) * FRACUNIT);
+						alpha = (int)(fixed_t(((double) alpha / (double) FRACUNIT) * ((double) MIN<fixed_t> (OPAQUE, Scale(harmor->Slots[armorType], OPAQUE, harmor->SlotsIncrement[armorType])) / (double) OPAQUE) * FRACUNIT));
 						texture = statusBar->Images[image];
 					}
 					else
@@ -663,7 +663,7 @@ class CommandDrawSwitchableImage : public CommandDrawImage
 
 				// Since we're not going to call our parent's tick() method,
 				// be sure to set the alpha value properly.
-				alpha = FRACUNIT;
+				alpha = (int)(FRACUNIT);
 				return;
 			}
 			CommandDrawImage::Tick(block, statusBar, hudChanged);
@@ -1695,7 +1695,7 @@ class CommandDrawSelectedInventory : public CommandDrawImage, private CommandDra
 					if(itemflash && itemflashFade)
 					{
 						fixed_t flashAlpha = fixed_t(((double) block->Alpha() / (double) FRACUNIT) * ((double) itemflashFade / (double) OPAQUE) * FRACUNIT);
-						statusBar->DrawGraphic(statusBar->Images[statusBar->invBarOffset + imgCURSOR], imgx-4, imgy+2, block->XOffset(), block->YOffset(), flashAlpha, block->FullScreenOffsets(),
+						statusBar->DrawGraphic(statusBar->Images[statusBar->invBarOffset + imgCURSOR], imgx-4, imgy+2, block->XOffset(), block->YOffset(), (int)(flashAlpha), block->FullScreenOffsets(),
 							translatable, false, offset);
 					}
 					CommandDrawImage::Draw(block, statusBar);
@@ -2211,7 +2211,7 @@ class CommandDrawInventoryBar : public SBarInfoCommand
 		
 			int bgalpha = block->Alpha();
 			if(translucent)
-				bgalpha = fixed_t((((double) block->Alpha() / (double) FRACUNIT) * ((double) HX_SHADOW / (double) FRACUNIT)) * FRACUNIT);
+				bgalpha = (int)(fixed_t((((double) block->Alpha() / (double) FRACUNIT) * ((double) HX_SHADOW / (double) FRACUNIT)) * FRACUNIT));
 		
 			AInventory *item;
 			unsigned int i = 0;
@@ -3616,7 +3616,7 @@ class CommandAlpha : public SBarInfoMainBlock
 		void	Parse(FScanner &sc, bool fullScreenOffsets)
 		{
 			sc.MustGetToken(TK_FloatConst);
-			alpha = fixed_t(FRACUNIT * sc.Float);
+			alpha = (int)(fixed_t(FRACUNIT * sc.Float));
 
 			// We don't want to allow all the options of a regular main block
 			// so skip to the SBarInfoCommandFlowControl.
